feat: Add training handler file, class and change on the code
