+++ src/crewai/agent.py
@@ -1,6 +1,6 @@
-from copy import deepcopy
 import os
 import uuid
+from copy import deepcopy
 from typing import Any, Dict, List, Optional, Tuple
 
 from langchain.agents.agent import RunnableAgent
@@ -25,8 +25,8 @@ from crewai.agents import CacheHandler, CrewAgentExecutor, CrewAgentParser, Tool
 from crewai.memory.contextual.contextual_memory import ContextualMemory
 from crewai.utilities import I18N, Logger, Prompts, RPMController
 from crewai.utilities.constants import TRAINED_AGENTS_DATA_FILE, TRAINING_DATA_FILE
-from crewai.utilities.fileHandler import PickleHandler
 from crewai.utilities.token_counter_callback import TokenCalcHandler, TokenProcess
+from crewai.utilities.training_handler import CrewTrainingHandler
 
 
 class Agent(BaseModel):
@@ -416,7 +416,7 @@ class Agent(BaseModel):
 
     def _training_handler(self, task_prompt: str) -> str:
         """Handle training data for the agent task prompt to improve output on Training."""
-        if data := PickleHandler(TRAINING_DATA_FILE).load():
+        if data := CrewTrainingHandler(TRAINING_DATA_FILE).load():
             agent_id = str(self.id)
 
             if data.get(agent_id):
@@ -431,7 +431,7 @@ class Agent(BaseModel):
 
     def _use_trained_data(self, task_prompt: str) -> str:
         """Use trained data for the agent task prompt to improve output."""
-        if data := PickleHandler(TRAINED_AGENTS_DATA_FILE).load():
+        if data := CrewTrainingHandler(TRAINED_AGENTS_DATA_FILE).load():
             if trained_data_output := data.get(self.role):
                 task_prompt += "You MUST follow these feedbacks: \n " + "\n - ".join(
                     trained_data_output["suggestions"]
